finish updating Command docs
